pyproject.toml
[tool.poetry]
name = "multi-agent-product-system"
version = "1.0.0"
description = "A comprehensive multi-agent system for automated product listing generation"
authors = ["AI Development Team <dev@company.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = ">=2.0.0"
asyncio-throttle = ">=1.0.2"
aiohttp = ">=3.8.0"
replicate = ">=0.15.0"
openai = ">=1.0.0"
Pillow = ">=9.0.0"
beautifulsoup4 = ">=4.11.0"
lxml = ">=4.9.0"
requests = ">=2.28.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.8.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"
pytest-mock = ">=3.10.0"
black = ">=22.0.0"
isort = ">=5.10.0"
mypy = ">=1.0.0"
gunicorn = ">=20.1.0"
uvicorn = {extras = ["standard"], version = ">=0.20.0"}
fastapi = ">=0.95.0"

[tool.poetry.scripts]
product-listing-system = "multi_agent_product_system.examples.usage_examples:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
